# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Employee:
    properties:
      address:
        type: string
      designation:
        type: string
      employeeId:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample Employee CRUD API.
  title: Employee API
  version: "1.0"
paths:
  /employees:
    get:
      description: Get a list of all employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
      summary: Get all employees
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Create a new employee with the input payload
      parameters:
      - description: Create employee
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
      summary: Create a new employee
      tags:
      - employees
  /employees/{id}:
    delete:
      description: Delete an employee by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an employee
      tags:
      - employees
    get:
      description: Get an employee by ID
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
      summary: Get an employee
      tags:
      - employees
    put:
      consumes:
      - application/json
      description: Update an employee with the input payload
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update employee
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
      summary: Update an employee
      tags:
      - employees
swagger: "2.0"